+++ Changelog.md
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+### 1.0.1 (20.02.2014)
+
+* symfony 2.3 LTS composer bugfix
+* vertical barcode rendering for image based barcodes implemented
+* minor changes in unit test code structure
+
 ### 1.0.0 (2014-02-17)
 
 * unit tests